# PERF: constructing string Series